% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_column_names.R
\name{combine_column_names}
\alias{combine_column_names}
\title{Combine column names in a data frame}
\usage{
combine_column_names(
  df,
  name_row_count,
  retain_original_names = TRUE,
  slide_headers = FALSE,
  separator = "_"
)
}
\arguments{
\item{df}{A data frame whose column names need to be combined.}

\item{name_row_count}{Integer, the number of rows in which column names
are broken.}

\item{retain_original_names}{Logical, if TRUE, retains the original names in
the combined name.}

\item{slide_headers}{Logical, if TRUE, slides the headers down to fill NAs.}

\item{separator}{Character, the character(s) to use when joining names.}
}
\value{
A data frame with combined column names.
}
\description{
This function combines column names in a data frame where the names are
broken across multiple rows. It reshapes the data and joins the names,
which allows for better organization and cleaner data. This function was
developed to tackle the challenges encountered when working with large Excel
files that contain many spanning headers. It is designed to be used in
conjunction with other functions in the epiCleanr package, like
`read_and_combine_sheet()`, but can also be used on it's own.
}
\examples{
\dontrun{
df <- mtcars
combine_column_names(df,
  name_row_count = 2, retain_original_names = TRUE,
  slide_headers = TRUE, separator = "_"
)
}
}
\seealso{
\code{\link[epiCleanr]{read_and_combine_sheet}} for a function
that uses this one.
}
