% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\docType{data}
\name{fake_epi_df_togo}
\alias{fake_epi_df_togo}
\title{Simulated Epidemiological Dataset for Togo}
\format{
A data frame with 900 rows and 8 variables:
\describe{
  \item{\code{country}}{The country of the data. Always 'Togo' in this
  dataset.}
  \item{\code{district}}{The state in Togo where the data was collected.
  Includes 15 different districts}
\item{\code{region}}{The region in Togo where the data was collected.
  Includes 4 different regions}
  \item{\code{month}}{The month when the data was recorded, ranging from
  1 to 12.}
  \item{\code{year}}{The year when the data was recorded, ranging from 2018
  to 2022.}
  \item{\code{malaria_tests}}{The number of tests conducted for malaria.
  Includes random missing values and outliers.}
  \item{\code{malaria_cases}}{The number of confirmed cases for malaria.
  Includes random missing values and outliers.}
  \item{\code{cholera_tests}}{The number of tests conducted for cholera.
  Includes random missing values and outliers.}
  \item{\code{cholera_cases}}{The number of confirmed cases for cholera.
  Includes random missing values and outliers.}
}
}
\usage{
data("fake_epi_df_togo")
}
\description{
A dataset representing simulated epidemiological data
  for malaria and cholera across various states in Togo from 2018-2022.
}
\details{
The dataset was generated using various steps:
\itemize{
  \item The dataset was initiated using `dplyr::expand_grid()` to create a
  combination of districts, region, months, and years.
  \item Random values were generated for \code{malaria_tests},
  \code{malaria_cases}, \code{cholera_tests}, and \code{cholera_cases} using
  a normal distribution.
  \item Outliers were introduced for 5% of the data in \code{cholera_tests},
  \code{malaria_cases}, and \code{cholera_cases}.
  \item 24% of the data in all 'tests' and 'cases' columns was randomly set
  to \code{NA}.
  \item The number of 'cases' was adjusted to make sure they are less than
  the number of 'tests'.
  \item Any negative 'cases' values were converted to their absolute values.
}
}
\examples{
\dontrun{
data('fake_epi_df_togo')
head(fake_epi_df_togo)
}

}
\keyword{datasets}
