% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_combine_sheet.R
\name{read_and_combine_sheet}
\alias{read_and_combine_sheet}
\title{Read an Excel sheet, clean and reshape its data}
\usage{
read_and_combine_sheet(excel_path, id_cols, sheet, rows)
}
\arguments{
\item{excel_path}{Character string representing the file path to the Excel
file.}

\item{id_cols}{Character vector representing the column names to be treated
as identifiers in the reshaping process.}

\item{sheet}{Character string representing the name of the Excel sheet to
be read.}

\item{rows}{Integer, the number of rows where the column names are broken.}
}
\value{
A tidied data frame with combined column names and reshaped data.
}
\description{
This function reads an Excel sheet, applies cleaning to the data, combines
column names, and reshapes the data into a more manageable format. This
function makes use of two other functions  in the epiCleanr package:
clean_names() and `combine_column_names()`.
}
\examples{
\dontrun{
# assuming that you have an Excel file 'data.xlsx' and the sheet 'Sheet1'
# has column names broken across the first 3 rows
read_and_combine_sheet("data.xlsx", c("id", "year"), "Sheet1", 3)
}
}
\seealso{
\code{\link[epiCleanr]{clean_names}},
\code{\link[epiCleanr]{combine_column_names}}
for functions used within this one.
}
