% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/togo_admin_df.R
\docType{data}
\name{togo_admin_df}
\alias{togo_admin_df}
\title{Togo Administrative Regions and Districts Dataset}
\format{
A data frame with variables:
\describe{
  \item{region}{The administrative level 1 division in Togo, also known as
  regions.}
  \item{district}{The administrative level 2 division in Togo, also known
  as districts.}
}
}
\usage{
data("togo_admin_df")
}
\description{
This dataset contains administrative level 1 and level 2 information for
Togo. It is intended for testing the functionalities of the epiCleanr package,
particularly those that relate to geographical subdivisions.
}
\examples{
\dontrun{
data("togo_admin_df")
head(togo_admin_df)
}

}
\keyword{datasets}
