% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean variable names}
\usage{
clean_names(var)
}
\arguments{
\item{var}{Character vector representing the variable name(s) to be cleaned.}
}
\value{
The cleaned variable name(s) as a character vector.
}
\description{
This function transforms a variable name into a standard, cleaned format.
The transformation includes converting to lower case, replacing spaces and
punctuation with underscores, replacing uppercase characters following
lowercase characters (camel case) with lowercase characters separated by an
underscore, and removing leading or trailing underscores.  The function is
inspired by the `clean_names` function from the janitor package and is
designed to be used in conjunction with the `combine_column_names` function
of the epiCleanr package.
}
\examples{
\dontrun{
clean_names("My variable NAME")
clean_names(c("var1", "Another Variable", "SOME_VAR"))
}
}
\seealso{
\preformatted{\link[janitor]{clean_names}} for the function that inspired
this one.
\preformatted{\link[epiCleanr]{combine_column_names}} for the function that uses this one.
}
