% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency_check.R
\name{consistency_check}
\alias{consistency_check}
\title{Consistency Check Function}
\usage{
consistency_check(data, tests, cases)
}
\arguments{
\item{data}{A data frame containing the test and case data.}

\item{tests}{A character vector specifying the column names for the test
data.}

\item{cases}{A character vector specifying the column names for the case
data. The length of 'tests' and 'cases' must be the same, and each element
in 'tests' corresponds to an element in 'cases'.}
}
\value{
A ggplot2::ggplot object showing the consistency between the number
of tests and cases. The x-axis represents the cases, and the y-axis
represents the tests. Each facet represents a disease, and the diagonal
line shows where the number of tests equals the number of cases.
}
\description{
This function performs a consistency check to ensure that the number of tests
is greater than the number of cases for given columns in a dataset. It
returns a ggplot2::ggplot object visualizing the results.
}
\examples{
\dontrun{
dummy_data <- data.frame(
      malaria_rdt_test = rnorm(1000, mean = 0, sd = 50),
      malaria_micro_test = rnorm(1000, mean = 0, sd = 10),
      malaria_rdt_cases = rnorm(1000, mean = 0, sd = 50),
      malaria_micro_cases = rnorm(1000, mean = 1, sd = 10))

consistency_check(dummy_data,
              tests = c("malaria_rdt_test","malaria_micro_test"),
              cases = c("malaria_rdt_cases", "malaria_micro_cases"))
}

}
