% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test.R
\name{create_test}
\alias{create_test}
\title{Create Test Function}
\usage{
create_test(
  dimension_test = NULL,
  combinations_test = NULL,
  row_duplicates = F,
  col_duplicates = F,
  min_threshold_test = NULL,
  max_threshold_test = NULL
)
}
\arguments{
\item{dimension_test}{A vector of two integers specifying the expected number
of rows and columns.}

\item{combinations_test}{A list with the elements `variables` (character
vector of variable names) and `expectation` (integer specifying the expected
number of unique combinations for each column).}

\item{row_duplicates}{Logical. If TRUE, checks for duplicate rows.}

\item{col_duplicates}{Logical. If TRUE, checks for duplicate columns.}

\item{min_threshold_test}{Named list of minimum threshold values for
specified columns.}

\item{max_threshold_test}{Named list of maximum threshold values for
specified columns.}
}
\value{
A function to be applied to the dataset.
}
\description{
This function creates a test function to perform various data validation
checks. The returned function can be applied to a dataset to perform the
specified tests.
}
\examples{
\dontrun{
my_tests <- create_test(
  dimension_test = c(14400, 3),
  combinations_test = list(
    variables = c("month", "year", "state"),
    expectation = 13 * 24 * 51
  ),
  row_duplicates = TRUE, col_duplicates = TRUE,
  max_threshold_test = list(polio_tests = 12),
  min_threshold_test = list(polio_tests = 221)
)
result <- my_tests(fake_epi_data)
}
}
