% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import Data from Various File Formats}
\usage{
import(file_path, ...)
}
\arguments{
\item{file_path}{Character string specifying the path to the input file or
a URL pointing to the dataset.}

\item{...}{Additional arguments to be passed to the underlying read
functions. These arguments are specific to the file format being imported.
Please refer to the documentation of each package used for more
information.}
}
\value{
A data frame or appropriate R object containing the imported data.
}
\description{
This function provides a unified interface for importing data from various
file formats supported by the `rio` package. The format is automatically
detected from the file extension to simplify the importing process.
}
\examples{
\dontrun{
# Import a CSV file
data_csv <- import("path/to/your/file.csv")

# Import an Excel file
data_excel <- import("path/to/your/file.xlsx")

# Import a Stata DTA file
data_dta <- import("path/to/your/file.dta")

# Import an RDS file
data_rds <- import("path/to/your/file.rds")

# Import an RData file
data_rdata <- import("path/to/your/file.RData")

# Import a JSON file
data_json <- import("path/to/your/file.json")

# Import an SPSS file
data_spss <- import("path/to/your/file.sav")

# Import a data.table file
data_dt <- import("path/to/your/file.dt")

# Import a dataset from a website (e.g., GitHub)
data_web <- import("https://github.com/your_username/your_dataset.xlsx")
}

}
\seealso{
\code{\link[rio:import]{foo::import()}},  which this function
is based on.
}
