% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_plot.R
\name{missing_plot}
\alias{missing_plot}
\title{Plot Missing data over time}
\usage{
missing_plot(data, x_var, y_var = NULL, miss_vars = NULL, use_rep_rate = FALSE)
}
\arguments{
\item{data}{A data frame containing the data to be visualized. Must include
columns specified in 'x_var', 'y_var', and 'vars'.}

\item{x_var}{A character string specifying the time variable in 'data'
(e.g., "year", "month"). Must be provided.}

\item{y_var}{An optional character string specifying the grouping
variable in 'data' (e.g., "state"). If provided, only one variable can be
specified in 'vars'.}

\item{use_rep_rate}{A logical value. If TRUE, the reporting rate is
visualized; otherwise, the proportion of missing data is visualized.
Defaults to FALSE}

\item{miss_value}{An optional character vector specifying the variables
to be visualized in 'data'. If NULL, all variables except 'x_var' and
'y_var' will be used.}
}
\value{
A ggplot2 object representing the tile plot.
}
\description{
This function visualizes the proportion of missing data or reporting rate for
 specified variables in a dataset. It creates a tile plot using ggplot2 where
 the x-axis represents time (e.g., year, month), and the y-axis represents
 either variables or groupings (e.g., state). The output can further be
 manipulated to one's needs.
}
\examples{
\dontrun{
# Create a data frame with all combinations of states and years
fake_data <- expand_grid(
  state = state.name,
  month = 1:12, year = 2000:2023
) |>
  mutate(
    measles = sample(0:1000, size = n(), replace = TRUE),
    polio = sample(0:500, size = n(), replace = TRUE),
    malaria = sample(0:300, size = n(), replace = TRUE),
    cholera = sample(0:700, size = n(), replace = TRUE)
  ) |>
  # Randomly set 10\% of the data in each disease column to NA
  mutate(across(
    c(measles, polio, malaria, cholera),
    ~ replace(., sample(1:n(), size = n() * 0.4), NA)
  ))

# Example function usage
missing_plot(fake_data,
  x_var = "year",
  miss_value = "polio",
  y_var = "state",
  use_rep_rate = T
)
}
}
